Add driver based Neo4jHealthIndicator.
This adds Neo4jNativeHealthIndicator that is based on the Neo4j Java Driver when a org.neo4j.driver.v1.Driver bean is present. It is preferred over the OGM based health indicator.

This works well with the new feature that allows the user to be responsible of the driver bean.